Extract OpaqueNodeMethods to own file
